Move formatters into separate file
